requirements.txt
requests
lxml
bs4
Pillow